+++ test/Confluent.Kafka.UnitTests/Serialization/Double.cs
@@ -28,7 +28,7 @@ namespace Confluent.Kafka.UnitTests.Serialization
         {
             foreach (var value in TestData)
             {
-                Assert.Equal(value, Deserializers.Double(SimpleSerializers.Double(value), false));
+                Assert.Equal(value, Deserializers.Double(Serializers.Double(value), false));
             }
         }
 
@@ -37,7 +37,7 @@ namespace Confluent.Kafka.UnitTests.Serialization
         {
             var buffer = new byte[] { 23, 0, 0, 0, 0, 0, 0, 0 };
             var value = BitConverter.ToDouble(buffer, 0);
-            var data = SimpleSerializers.Double(value);
+            var data = Serializers.Double(value);
             Assert.Equal(23, data[7]);
             Assert.Equal(0, data[0]);
         }